requirements.txt
numpy>=2.0.0
pandas==2.2.3
plotly==5.24.1
groq==0.15.0
httpx==0.27.0
pyyaml==6.0.2
python-dotenv==1.0.1
pyvis==0.3.2
python-dateutil>=2.8.2
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
